+++ RELEASE_NOTES.md
@@ -54,6 +54,10 @@ term.
 * ADDED: Also search for settings files on the `%APPDATA%\Logtalk` directory
 when running on Windows operating-systems.
 
+* ADDED: Experimental `@/1` goal annotation to allow calling from a
+complementing category a predicate definition in the patched object
+that is being replaced by a new definition in the category.
+
 * IMPROVED: Detect more cases of unification goals that are always true or
 always false.